requirements.txt
attrs==23.1.0
boxsdk==3.7.1
certifi==2023.5.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.30.0
requests-toolbelt==0.10.1
six==1.16.0
tzdata==2023.3
urllib3==1.26.15